Adds db update command.
This implements the first version of the db update command. Now, any
changes in the database can be passed on to end users in the system
itself.

All the stuff inside migrations is generated by alembic. The
alembic.ini has instructions on alembic config. This package and its
dependencies are also added into the requirements.txt file.